[FIX] resp is not None by definition, so we check the status_code
request.Response() is not None, but the related status_code
is. This is why we would get a traceback for resp.ok or resp
which is redirected to resp.ok which is checking the code for
an integer.

Also make sure that too many retries will result in a ReadTimeout